+++ src/IbanNet/IbanParser.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
+using IbanNet.Registry;
 
 namespace IbanNet
 {
@@ -10,6 +11,23 @@ namespace IbanNet
     public sealed class IbanParser : IIbanParser
     {
         private readonly IIbanValidator _ibanValidator;
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="IbanParser" /> class using specified <paramref name="registry" />.
+        /// </summary>
+        /// <param name="registry">The registry.</param>
+        public IbanParser(IIbanRegistry registry)
+        {
+            if (registry is null)
+            {
+                throw new ArgumentNullException(nameof(registry));
+            }
+
+            _ibanValidator = new IbanValidator(new IbanValidatorOptions
+            {
+                Registry = registry
+            });
+        }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="IbanParser" /> class using specified <paramref name="ibanValidator" />.
@@ -60,13 +78,9 @@ namespace IbanNet
             iban = null;
             exceptionThrown = null;
 
-            // Although our validator normalizes too, we can't rely on this fact if other implementations
-            // are provided (like mocks, or maybe faster validators). Thus, to ensure this class correctly
-            // represents the IBAN value, we normalize inline here and take the penalty.
-            string? normalizedValue = Iban.NormalizeOrNull(value);
             try
             {
-                validationResult = _ibanValidator.Validate(normalizedValue);
+                validationResult = _ibanValidator.Validate(value);
             }
             catch (Exception ex)
             {
@@ -80,7 +94,7 @@ namespace IbanNet
                 return false;
             }
 
-            iban = new Iban(normalizedValue!);
+            iban = new Iban(validationResult.AttemptedValue!);
             return true;
         }
     }